cast for lvalue
git-svn-id: file:///Users/walter/linux/walter/src/svnroot/trunk/bbs100@126 d90a1141-1c03-0410-82d0-813645769019

+++ src/include/FileFormat.h
@@ -78,9 +78,9 @@
 
 #define FF1_LOAD_ULONG(x,y)		if (!strcmp(buf, (x))) {	\
 		if (!*p)											\
-			(unsigned long)(y) = 0UL;						\
+			(y) = 0UL;										\
 		else												\
-			(unsigned long)(y) = strtoul(p, NULL, 10);		\
+			(y) = strtoul(p, NULL, 10);						\
 		continue;											\
 	}
